Changing design

+++ WebServerr/src/servlet/main/page/hitchhiker/CancelRideHitchhikerServlet.java
@@ -1,5 +1,7 @@
 package servlet.main.page.hitchhiker;
 
+import database.DriverRideDAO;
+import database.HitchhikerRideDAO;
 import database.MatchedDAO;
 import database.Users;
 import entity.ServerClient;
@@ -19,30 +21,35 @@ import java.util.HashMap;
 public class CancelRideHitchhikerServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String eventName = request.getParameter("eventName");
-
         String userName = (String)request.getSession().getAttribute("userName");
         ServerClient client = Users.getUserByUserName(userName);
 
         HashMap<String, ServerClient> webUsers = Users.getWebUsers();
 
-        //need to remove hitchhiker from the driver list and update the curr amount of hitchhikers
         HitchhikerRide hitchhikerRide = client.getHitchhikingEventByName(eventName);
-        if (hitchhikerRide != null)
-        {
-            String driverName = hitchhikerRide.getDriverName();
-            if (!driverName.isEmpty() && webUsers.containsKey(driverName)) {
+        String driverName = null;
+
+        if (hitchhikerRide != null) {
+            driverName = hitchhikerRide.getDriverName();
+            if (driverName != null && !driverName.isEmpty() && webUsers.containsKey(driverName)) {
                 ServerClient driver = webUsers.get(driverName);
                 DriverRide driverRide = driver.getDrivingEventByName(eventName);
                 MatchedDAO matchedDAO = new MatchedDAO();
-                matchedDAO.deleteMatch(eventName,driverName,hitchhikerRide.getHitchhikerUserName());
-                driverRide.removeHitchhikerByFullName(client.getUserName()); //remove hitchiker from the driverRide
-                Users.removeSpecificHitchhikerRide(hitchhikerRide.getHitchhikerUserName(),eventName);
-                //driverRide.lowerTotalFuelReturns(hitchhikerRide.getFuelMoney());
+                matchedDAO.deleteMatch(eventName, driverName, hitchhikerRide.getHitchhikerUserName());
+                driverRide.removeHitchhikerByFullName(client.getUserName());
+                Users.removeSpecificHitchhikerRide(hitchhikerRide.getHitchhikerUserName(), eventName);
             }
         }
+
         client.getHitchhikingEvents().remove(eventName);
 
+        DriverRideDAO driverRideDAO = new DriverRideDAO();
+        boolean isMatchFound = false;
+        if (driverName != null) {
+            isMatchFound = driverRideDAO.matchHitchhikersToDriver(driverName, eventName);
+        }
 
         response.sendRedirect("hitchhikerOptionsMenu.jsp?userName="+userName);
+
     }
 }
